set authenticated require for relayer api

+++ backend/api/relayer.py
@@ -1,47 +1,58 @@
 from playhouse.shortcuts import model_to_dict
 from peewee import ProgrammingError
 from model import Relayer
-from exception import InvalidValueException, MissingArgumentException
+from exception import InvalidValueException, MissingArgumentException, UserAuthorizationException
+from util.decorator import authenticated
 from .base import BaseHandler
 
 
 class RelayerHandler(BaseHandler):
 
-    async def get(self):
-        relayers = []
-        relayers = [model_to_dict(relayer or {}) for relayer in Relayer.select()]
+    @authenticated
+    async def get(self, user):
+        relayers = [model_to_dict(relayer or {}) for relayer in Relayer.select().where(Relayer.owner == user)]
         self.json_response(relayers)
 
-    async def post(self):
+    @authenticated
+    async def post(self, user):
         """Add new relayer"""
         relayer = self.request_body
+
+        if user != relayer['owner']:
+            raise InvalidValueException('Owner address does not match user_address')
+
         try:
             obj = await self.application.objects.create(Relayer, **relayer)
             self.json_response(model_to_dict(obj))
         except Exception:
             raise InvalidValueException('relayer payload is invalid: {param}'.format(param=str(relayer)))
 
-    async def patch(self):
+    @authenticated
+    async def patch(self, user):
         """Update existing relayer"""
         relayer = self.request_body
         relayer_id = relayer.get('id', None)
+        relayer_owner = relayer.get('owner', None)
 
         if not relayer_id:
             raise MissingArgumentException('missing relayer id')
+
+        if not relayer_owner or relayer_owner != user:
+            raise InvalidValueException('Owner address does not match user_address')
 
         del relayer['id']
 
         try:
             query = (Relayer.update(**relayer).where(Relayer.id == relayer_id).returning(Relayer))
             cursor = query.execute()
-            obj = cursor[0]
             self.json_response(model_to_dict(obj))
         except IndexError:
             raise InvalidValueException('relayer id={param} does not exist'.format(param=str(relayer_id)))
         except ProgrammingError:
             raise InvalidValueException('update payload is invalid: {param}'.format(param=str(relayer)))
 
-    async def delete(self):
+    @authenticated
+    async def delete(self, user):
         """Delete a relayer"""
         relayer_id = self.get_argument('id', None)
 
@@ -50,7 +61,15 @@ class RelayerHandler(BaseHandler):
 
         try:
             relayer = Relayer.get(Relayer.id == relayer_id)
+
+            if relayer.owner != user:
+                raise InvalidValueException
+
             relayer.delete_instance()
             self.json_response({})
+
+        except InvalidValueException:
+            raise InvalidValueException('Owner address does not match user_address')
+
         except Exception:
             raise InvalidValueException('invalid relayer id: relayer with id={} does not exist'.format(relayer_id))
